backend/test_phase2.py
"""
Phase 2 Testing Script

Tests MockEngine and HFAPIEngine, verifies all exit criteria.
"""
import httpx
import asyncio
import json
from sqlalchemy import select
from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
from sqlalchemy.orm import sessionmaker
import os
from dotenv import load_dotenv

# Load environment
load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")

async def test_mock_engine():
    """Test 1: MockEngine - Create and run experiment"""
    print("=" * 60)
    print("TEST 1: MockEngine Inference")
    print("=" * 60)
    
    base_url = "http://localhost:8000/api/v1"
    
    async with httpx.AsyncClient(timeout=30.0) as client:
        # Create experiment
        print("\n1. Creating experiment...")
        experiment_data = {
            "name": "Test Mock Inference",
            "description": "Testing MockEngine for Phase 2",
            "config": {
                "model_name": "microsoft/phi-2",
                "reasoning_method": "naive",
                "dataset_name": "triviaqa",
                "generation_params": {
                    "max_tokens": 256,
                    "temperature": 0.7,
                    "top_p": 0.9
                },
                "seed": 42
            }
        }
        
        response = await client.post(f"{base_url}/experiments/", json=experiment_data)
        if response.status_code != 201:
            print(f"❌ Failed to create experiment: {response.status_code}")
            print(response.text)
            return None
        
        experiment = response.json()
        experiment_id = experiment["id"]
        print(f"✅ Created experiment: {experiment_id}")
        print(f"   Status: {experiment['status']}")
        
        # Run experiment
        print(f"\n2. Running experiment...")
        response = await client.post(f"{base_url}/experiments/{experiment_id}/run")
        if response.status_code != 200:
            print(f"❌ Failed to run experiment: {response.status_code}")
            print(response.text)
            return None
        
        print(f"✅ Experiment started")
        
        # Poll for completion
        print("\n3. Waiting for completion...")
        for i in range(20):  # Wait up to 20 seconds
            await asyncio.sleep(1)
            response = await client.get(f"{base_url}/experiments/{experiment_id}")
            exp = response.json()
            status = exp["status"]
            print(f"   [{i+1}s] Status: {status}")
            
            if status == "completed":
                print(f"✅ Experiment completed!")
                return experiment_id
            elif status == "failed":
                print(f"❌ Experiment failed: {exp.get('error_message', 'Unknown error')}")
                return None
        
        print("❌ Timeout waiting for completion")
        return None

async def verify_database(experiment_id):
    """Verify database has all required data"""
    print("\n" + "=" * 60)
    print("DATABASE VERIFICATION")
    print("=" * 60)
    
    engine = create_async_engine(DATABASE_URL)
    async_session = sessionmaker(engine, class_=AsyncSession, expire_on_commit=False)
    
    async with async_session() as session:
        # Import models
        import sys
        sys.path.insert(0, "C:/Users/FAZLUL/Desktop/MainProject/LlmForge/backend")
        from app.models.run import Run
        
        # Get all runs for this experiment
        result = await session.execute(
            select(Run).where(Run.experiment_id == experiment_id)
        )
        runs = result.scalars().all()
        
        print(f"\n✓ Total runs logged: {len(runs)}")
        
        if len(runs) == 0:
            print("❌ No runs found in database!")
            return False
        
        # Check exit criteria
        print("\n" + "-" * 60)
        print("EXIT CRITERIA VERIFICATION")
        print("-" * 60)
        
        criteria_passed = 0
        criteria_total = 5
        
        # Criterion 1: 10 runs logged
        if len(runs) == 10:
            print("✅ [1/5] Run 10 consecutive inferences without crashes")
            criteria_passed += 1
        else:
            print(f"❌ [1/5] Expected 10 runs, got {len(runs)}")
        
        # Criterion 2: Token counts populated
        token_counts_valid = all(r.tokens_input > 0 and r.tokens_output > 0 for r in runs)
        if token_counts_valid:
            print("✅ [2/5] Token counts match expected (input + output)")
            criteria_passed += 1
        else:
            print("❌ [2/5] Some token counts are 0 or missing")
        
        # Criterion 3: Latency reasonable
        avg_latency = sum(r.latency_ms for r in runs) / len(runs) if runs else 0
        if avg_latency < 5000:  # <5s
            print(f"✅ [3/5] Latency is reasonable ({avg_latency:.0f}ms avg < 5000ms)")
            criteria_passed += 1
        else:
            print(f"❌ [3/5] Latency too high ({avg_latency:.0f}ms avg)")
        
        # Criterion 4: All non-null required fields
        all_fields_valid = all(
            r.input_text and r.output_text and 
            r.tokens_input is not None and r.tokens_output is not None and
            r.latency_ms is not None
            for r in runs
        )
        if all_fields_valid:
            print("✅ [4/5] Runs table has all non-null required fields")
            criteria_passed += 1
        else:
            print("❌ [4/5] Some required fields are null")
        
        # For MockEngine, we'll mark error handling as pass (will test with HF API)
        print("⏭️  [5/5] API error handling (will test with HF API)")
        
        # Show sample runs
        print("\n" + "-" * 60)
        print("SAMPLE RUNS")
        print("-" * 60)
        for i, run in enumerate(runs[:3], 1):
            print(f"\n{i}. Example ID: {run.example_id}")
            print(f"   Input: {run.input_text[:50]}...")
            print(f"   Output: {run.output_text[:50]}...")
            print(f"   Tokens: {run.tokens_input} → {run.tokens_output}")
            print(f"   Latency: {run.latency_ms:.0f}ms")
            print(f"   Correct: {run.is_correct}")
        
        print("\n" + "=" * 60)
        print(f"PHASE 2 PROGRESS: {criteria_passed}/4 criteria passed (MockEngine)")
        print("=" * 60)
        
        return criteria_passed >= 4

async def main():
    print("\n🧪 PHASE 2 EXIT CRITERIA TESTING\n")
    
    # Test MockEngine
    experiment_id = await test_mock_engine()
    
    if not experiment_id:
        print("\n❌ MockEngine test failed!")
        return
    
    # Verify database
    success = await verify_database(experiment_id)
    
    if success:
        print("\n✅ PHASE 2 TESTS PASSED!")
        print("\nNext: Test with HuggingFace API by setting INFERENCE_ENGINE=hf_api")
    else:
        print("\n❌ Some exit criteria not met")

if __name__ == "__main__":
    asyncio.run(main())
