backend/requirements.txt
# LLM Research Platform - Backend

## Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

## Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

## ML & Inference
# TODO (Iteration 1): Uncomment when ready to run models
# transformers>=4.36.0
# torch>=2.1.0
# accelerate>=0.25.0
# bitsandbytes>=0.41.0  # For 4-bit quantization

## ML & Inference (NO PyTorch!)
# huggingface-hub: Lightweight API client for HF Inference API
huggingface-hub>=0.20.0

## Evaluation
# TODO (Iteration 2): Uncomment for faithfulness metrics
# numpy>=1.26.0
# scipy>=1.11.0

## Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.0
